Refactor: move the MoveButtonGroup to the widgets.
